requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pyyaml
jinja2
